ml-service/requirements.txt
flask
flask-cors
google-generativeai
PyPDF2
pdfplumber
numpy
pandas
scikit-learn
joblib
python-dotenv
pytesseract
opencv-python
Pillow
pdf2image
python-magic
cryptography
